requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
google-api-python-client==2.86.0
openai==1.12.0
gunicorn==21.2.0
requests==2.31.0
feedparser==6.0.10 

